+++ client/CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file.
 
 - Change field component's default text color to #565555
 - Implement table component for all cards with table in dashboard page
+- Apply active state for nested navigation of sidebar
 - Minor adjustment on initial data seed of deparmtent
 - Minor adjustment of the department's spacing