+++ client/CHANGELOG.md
@@ -6,12 +6,21 @@ All notable changes to this project will be documented in this file.
 
 ### 🚀 Added
 
+### 🔄 Changed
+
+### 🐛 Fixed
+
+### 🗑️ Removed
+
+## [1.2.0] - 2026-03-10
+
+### 🚀 Added
+
 - New feature (service module) with basic functionality and initial pages (list, create, update & view)
 - Add department column on position table
 - Add another filter on position table (by department)
 
 ### 🔄 Changed
-
 
 ### 🐛 Fixed